requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.8.1
google-generativeai==0.8.5
python-dotenv==1.1.0
motor==3.7.0
pymongo==4.11.3
